auto_trading/requirements.txt
# 키움 API 연동
PyQt5==5.15.9

# 데이터 처리
pandas==2.0.3
numpy==1.25.2

# 환경 변수 관리
python-dotenv==1.0.0

# 로깅
loguru==0.7.0

# 날짜/시간 처리
python-dateutil==2.8.2

# 데이터 저장 (선택사항)
openpyxl==3.1.2

